+++ app/src/main/java/com/doilio/shopifymemory/model/Options.kt
@@ -4,8 +4,8 @@ import com.google.gson.annotations.SerializedName
 
 data class Options (
 
-    @SerializedName("id") val id : Int,
-    @SerializedName("product_id") val product_id : Int,
+    @SerializedName("id") val id : Long,
+    @SerializedName("product_id") val product_id : Long,
     @SerializedName("name") val name : String,
     @SerializedName("position") val position : Int,
     @SerializedName("values") val values : List<String>
